created utils function for popup and flytostore
