+++ utils/flyToStore.tsx
@@ -1,0 +1,6 @@
+export default function flyToStore(currentFeature: any, map: any) {
+  map.current.flyTo({
+    center: currentFeature.geometry.coordinates,
+    zoom: 15,
+  });
+}
